languages - move language associations closer to languages registry
