+++ src/vs/workbench/services/language/common/languageService.ts
@@ -3,9 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as nls from 'vs/nls';
-import * as mime from 'vs/base/common/mime';
-import * as resources from 'vs/base/common/resources';
+import { localize } from 'vs/nls';
+import { registerLanguageAssociation, clearLanguageAssociations } from 'vs/editor/common/services/languagesAssociations';
+import { joinPath } from 'vs/base/common/resources';
 import { URI } from 'vs/base/common/uri';
 import { ILanguageExtensionPoint, ILanguageService } from 'vs/editor/common/services/language';
 import { LanguageService } from 'vs/editor/common/services/languageService';
@@ -30,25 +30,25 @@ export interface IRawLanguageExtensionPoint {
 export const languagesExtPoint: IExtensionPoint<IRawLanguageExtensionPoint[]> = ExtensionsRegistry.registerExtensionPoint<IRawLanguageExtensionPoint[]>({
 	extensionPoint: 'languages',
 	jsonSchema: {
-		description: nls.localize('vscode.extension.contributes.languages', 'Contributes language declarations.'),
+		description: localize('vscode.extension.contributes.languages', 'Contributes language declarations.'),
 		type: 'array',
 		items: {
 			type: 'object',
 			defaultSnippets: [{ body: { id: '${1:languageId}', aliases: ['${2:label}'], extensions: ['${3:extension}'], configuration: './language-configuration.json' } }],
 			properties: {
 				id: {
-					description: nls.localize('vscode.extension.contributes.languages.id', 'ID of the language.'),
+					description: localize('vscode.extension.contributes.languages.id', 'ID of the language.'),
 					type: 'string'
 				},
 				aliases: {
-					description: nls.localize('vscode.extension.contributes.languages.aliases', 'Name aliases for the language.'),
+					description: localize('vscode.extension.contributes.languages.aliases', 'Name aliases for the language.'),
 					type: 'array',
 					items: {
 						type: 'string'
 					}
 				},
 				extensions: {
-					description: nls.localize('vscode.extension.contributes.languages.extensions', 'File extensions associated to the language.'),
+					description: localize('vscode.extension.contributes.languages.extensions', 'File extensions associated to the language.'),
 					default: ['.foo'],
 					type: 'array',
 					items: {
@@ -56,32 +56,32 @@ export const languagesExtPoint: IExtensionPoint<IRawLanguageExtensionPoint[]> = 
 					}
 				},
 				filenames: {
-					description: nls.localize('vscode.extension.contributes.languages.filenames', 'File names associated to the language.'),
+					description: localize('vscode.extension.contributes.languages.filenames', 'File names associated to the language.'),
 					type: 'array',
 					items: {
 						type: 'string'
 					}
 				},
 				filenamePatterns: {
-					description: nls.localize('vscode.extension.contributes.languages.filenamePatterns', 'File name glob patterns associated to the language.'),
+					description: localize('vscode.extension.contributes.languages.filenamePatterns', 'File name glob patterns associated to the language.'),
 					type: 'array',
 					items: {
 						type: 'string'
 					}
 				},
 				mimetypes: {
-					description: nls.localize('vscode.extension.contributes.languages.mimetypes', 'Mime types associated to the language.'),
+					description: localize('vscode.extension.contributes.languages.mimetypes', 'Mime types associated to the language.'),
 					type: 'array',
 					items: {
 						type: 'string'
 					}
 				},
 				firstLine: {
-					description: nls.localize('vscode.extension.contributes.languages.firstLine', 'A regular expression matching the first line of a file of the language.'),
+					description: localize('vscode.extension.contributes.languages.firstLine', 'A regular expression matching the first line of a file of the language.'),
 					type: 'string'
 				},
 				configuration: {
-					description: nls.localize('vscode.extension.contributes.languages.configuration', 'A relative path to a file containing configuration options for the language.'),
+					description: localize('vscode.extension.contributes.languages.configuration', 'A relative path to a file containing configuration options for the language.'),
 					type: 'string',
 					default: './language-configuration.json'
 				}
@@ -110,7 +110,7 @@ export class WorkbenchLanguageService extends LanguageService {
 				let extension = extensions[i];
 
 				if (!Array.isArray(extension.value)) {
-					extension.collector.error(nls.localize('invalid', "Invalid `contributes.{0}`. Expected an array.", languagesExtPoint.name));
+					extension.collector.error(localize('invalid', "Invalid `contributes.{0}`. Expected an array.", languagesExtPoint.name));
 					continue;
 				}
 
@@ -119,7 +119,7 @@ export class WorkbenchLanguageService extends LanguageService {
 					if (isValidLanguageExtensionPoint(ext, extension.collector)) {
 						let configuration: URI | undefined = undefined;
 						if (ext.configuration) {
-							configuration = resources.joinPath(extension.description.extensionLocation, ext.configuration);
+							configuration = joinPath(extension.description.extensionLocation, ext.configuration);
 						}
 						allValidLanguages.push({
 							id: ext.id,
@@ -158,7 +158,7 @@ export class WorkbenchLanguageService extends LanguageService {
 		const configuration = this._configurationService.getValue<IFilesConfiguration>();
 
 		// Clear user configured mime associations
-		mime.clearTextMimes(true /* user configured */);
+		clearLanguageAssociations(true /* user configured */);
 
 		// Register based on settings
 		if (configuration.files?.associations) {
@@ -166,7 +166,7 @@ export class WorkbenchLanguageService extends LanguageService {
 				const langId = configuration.files.associations[pattern];
 				const mimeType = this.getMimeType(langId) || `text/x-${langId}`;
 
-				mime.registerTextMime({ id: langId, mime: mimeType, filepattern: pattern, userConfigured: true });
+				registerLanguageAssociation({ id: langId, mime: mimeType, filepattern: pattern, userConfigured: true });
 			});
 		}
 
@@ -186,35 +186,35 @@ function isUndefinedOrStringArray(value: string[]): boolean {
 
 function isValidLanguageExtensionPoint(value: IRawLanguageExtensionPoint, collector: ExtensionMessageCollector): boolean {
 	if (!value) {
-		collector.error(nls.localize('invalid.empty', "Empty value for `contributes.{0}`", languagesExtPoint.name));
+		collector.error(localize('invalid.empty', "Empty value for `contributes.{0}`", languagesExtPoint.name));
 		return false;
 	}
 	if (typeof value.id !== 'string') {
-		collector.error(nls.localize('require.id', "property `{0}` is mandatory and must be of type `string`", 'id'));
+		collector.error(localize('require.id', "property `{0}` is mandatory and must be of type `string`", 'id'));
 		return false;
 	}
 	if (!isUndefinedOrStringArray(value.extensions)) {
-		collector.error(nls.localize('opt.extensions', "property `{0}` can be omitted and must be of type `string[]`", 'extensions'));
+		collector.error(localize('opt.extensions', "property `{0}` can be omitted and must be of type `string[]`", 'extensions'));
 		return false;
 	}
 	if (!isUndefinedOrStringArray(value.filenames)) {
-		collector.error(nls.localize('opt.filenames', "property `{0}` can be omitted and must be of type `string[]`", 'filenames'));
+		collector.error(localize('opt.filenames', "property `{0}` can be omitted and must be of type `string[]`", 'filenames'));
 		return false;
 	}
 	if (typeof value.firstLine !== 'undefined' && typeof value.firstLine !== 'string') {
-		collector.error(nls.localize('opt.firstLine', "property `{0}` can be omitted and must be of type `string`", 'firstLine'));
+		collector.error(localize('opt.firstLine', "property `{0}` can be omitted and must be of type `string`", 'firstLine'));
 		return false;
 	}
 	if (typeof value.configuration !== 'undefined' && typeof value.configuration !== 'string') {
-		collector.error(nls.localize('opt.configuration', "property `{0}` can be omitted and must be of type `string`", 'configuration'));
+		collector.error(localize('opt.configuration', "property `{0}` can be omitted and must be of type `string`", 'configuration'));
 		return false;
 	}
 	if (!isUndefinedOrStringArray(value.aliases)) {
-		collector.error(nls.localize('opt.aliases', "property `{0}` can be omitted and must be of type `string[]`", 'aliases'));
+		collector.error(localize('opt.aliases', "property `{0}` can be omitted and must be of type `string[]`", 'aliases'));
 		return false;
 	}
 	if (!isUndefinedOrStringArray(value.mimetypes)) {
-		collector.error(nls.localize('opt.mimetypes', "property `{0}` can be omitted and must be of type `string[]`", 'mimetypes'));
+		collector.error(localize('opt.mimetypes', "property `{0}` can be omitted and must be of type `string[]`", 'mimetypes'));
 		return false;
 	}
 	return true;
